refactor(evaluation): 평가 로직 개선을 위한 서비스 위임 및 질문 컨텍스트 유틸리티 추가
